# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/ip/query": {
            "get": {
                "description": "获取IP相关地理位置、ASN信息，综合判定威胁类型如:远控（C2）、傀儡机（Zombie）、失陷主机（Compromised）、扫描（Scanner）、钓鱼（Phishing）等，相关攻击团伙或安全事件标签，原始情报，相关样本信息等。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "可针对业务日志，以及从防火墙、WAF等安防设备中获取的外部IP，进行分析。",
                "operationId": "ip-query",
                "parameters": [
                    {
                        "type": "string",
                        "description": "resource string",
                        "name": "resource",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "exclude string",
                        "name": "exclude",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "lang string",
                        "name": "lang",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tb.Response"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "tb.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "response_code": {
                    "type": "integer"
                },
                "verbose_msg": {
                    "type": "string"
                }
            }
        }
    }
}